fix: _read 的地址转换经 uintptr_t，消除宿主机 -Wint-to-pointer-cast
base_addr 是"地址整数"（uint32_t 的绝对 Flash 地址），直接转 void* 在 64 位宿主机
测试（x86_64）上因宽度不足而告警；目标机（Cortex-M4，32 位）本就同宽、无此告警。
经 uintptr_t 中转既表达"这是地址"的意图，也在目标机上与原先完全等价（零 codegen 变化）。

测试侧本就以 MAP_32BIT 保证该地址落在 32 位可寻址区（见 test_iap_cfg.c:86），故不涉及截断。

至此自研层 warning 归零：make test 15 套全绿、两个固件目标 Build complete。

+++ Device/Storage/dev_flash_int.c
@@ -24,7 +24,9 @@ static int32_t _init(dev_storage_t *dev)
 static int32_t _read(dev_storage_t *dev, uint32_t addr, uint8_t *buf, uint32_t len)
 {
     dev_flash_int_t *self = container_of(dev, dev_flash_int_t, base);
-    memcpy(buf, (void *)(self->base_addr + addr), len);
+    /* base_addr 是"地址整数"：经 uintptr_t 转指针，既表达意图，也让 64 位宿主机测试
+     * 不再报 -Wint-to-pointer-cast（目标机 32 位下与原先完全等价，无 codegen 变化） */
+    memcpy(buf, (void *)(uintptr_t)(self->base_addr + addr), len);
     return 0; /* 约定: 0 = 成功（不返回字节数） */
 }
